pyproject.toml
[project]
name = "learn-graphics"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce==2.5.6",
]
